Probability of detection estimation when lot with homogeneous contaminations.

scenario_1_pd provides a probability of detection under scenario 1 of modelling the quantity of material sampled in the risk assessment study.

scenario_1_pd(mu, sd, m, type, n_sim)

Arguments

mu

the the mean concentration (\(\mu\)).

sd

standard deviation on the log10 scale (default value 0.8).

m

the vector of incremental samples (with equal/unequal weights).

type

what type of the results you would like to consider such as "theory" or "simulation".

n_sim

number of simulations (large simulations provide more precise estimation).

Value

Probability of detection when lot with homogeneous contaminants by using theoretical or simulation-based results.

Details

scenario_1_pd provides a probability of detection under scenario 1 of modelling the quantity of material sampled in the risk assessment study. Under this scenario (lot with homogeneous contaminations), we employed Poisson distribution to the model number of micro-organisms in the incremental samples. Based on the food safety literature, the expected cell count is given by \(\lambda = 10^{\mu+log(10)\sigma^2/2}\). (this section will be updated later on)

Examples

mu <- -3
sd <- 0.8
m <- c(5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5)
scenario_1_pd(mu, sd = 0.8, m, type = "theory")
#> [1] 0.4204717
scenario_1_pd(mu, sd = 0.8, m, type = "simulation", n_sim = 2000000)
#> [1] 0.4204645